# docs/swagger.yaml
basePath: /
definitions:
  AppError:
    description: Error that can be returned by the application
    properties:
      message:
        type: string
      statusCode:
        type: integer
    type: object
  Course:
    description: Model for a golf course
    properties:
      compact:
        type: boolean
      createdAt:
        description: '@Description Creation date of the model, automatically generated'
        example: 2021-01-01T00:00:00Z (auto-generated)
        type: string
      golf_id:
        type: string
      id:
        description: '@Description ID of the model, automatically generated'
        example: 123e4567-e89b-12d3-a456-426614174000 (auto-generated)
        type: string
      name:
        type: string
      num_holes:
        type: integer
      pitch_and_putt:
        type: boolean
      updatedAt:
        description: '@Description Last update date of the model, automatically generated'
        example: 2021-01-01T00:00:00Z (auto-generated)
        type: string
    type: object
  Golf:
    description: Model for a golf
    properties:
      city:
        description: '@Description City where the golf course is located'
        example: Guyancourt
        type: string
      createdAt:
        description: '@Description Creation date of the model, automatically generated'
        example: 2021-01-01T00:00:00Z (auto-generated)
        type: string
      googleMapLinks:
        description: '@Description Google Maps link for the golf course'
        example: https://maps.google.com/?q=Golf+National
        type: string
      id:
        description: '@Description ID of the model, automatically generated'
        example: 123e4567-e89b-12d3-a456-426614174000 (auto-generated)
        type: string
      latitude:
        description: '@Description Latitude of the golf course'
        example: 48.754
        type: number
      longitude:
        description: '@Description Longitude of the golf course'
        example: 2.074
        type: number
      name:
        description: '@Description Name of the golf course'
        example: Golf National
        type: string
      postalCode:
        description: '@Description Postal code of the golf course location'
        example: "78280"
        type: string
      updatedAt:
        description: '@Description Last update date of the model, automatically generated'
        example: 2021-01-01T00:00:00Z (auto-generated)
        type: string
    type: object
  Hole:
    description: Model for a hole
    properties:
      courseID:
        type: string
      createdAt:
        description: '@Description Creation date of the model, automatically generated'
        example: 2021-01-01T00:00:00Z (auto-generated)
        type: string
      holeNumber:
        type: integer
      id:
        description: '@Description ID of the model, automatically generated'
        example: 123e4567-e89b-12d3-a456-426614174000 (auto-generated)
        type: string
      par:
        type: integer
      updatedAt:
        description: '@Description Last update date of the model, automatically generated'
        example: 2021-01-01T00:00:00Z (auto-generated)
        type: string
    type: object
  Tee:
    description: Model for a tee
    properties:
      color:
        type: string
      createdAt:
        description: '@Description Creation date of the model, automatically generated'
        example: 2021-01-01T00:00:00Z (auto-generated)
        type: string
      distance:
        type: integer
      holeID:
        type: string
      id:
        description: '@Description ID of the model, automatically generated'
        example: 123e4567-e89b-12d3-a456-426614174000 (auto-generated)
        type: string
      updatedAt:
        description: '@Description Last update date of the model, automatically generated'
        example: 2021-01-01T00:00:00Z (auto-generated)
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This project is a comprehensive golf management system designed to
    facilitate the management of golf courses, including tracking of golf courses,
    holes, and tees.
  title: ProjectPAR API
  version: 1.0.0-Beta
paths:
  /v1/courses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Course'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Get all courses
      tags:
      - Courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Course
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/Course'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Create a new course
      tags:
      - Courses
  /v1/courses/{id}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Delete a course by ID
      tags:
      - Courses
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Get a course by ID
      tags:
      - Courses
    patch:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Course
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/Course'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Update a course by ID
      tags:
      - Courses
  /v1/golfs:
    get:
      description: Get all golfs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Golf'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Get all golfs
      tags:
      - Golfs
    post:
      consumes:
      - application/json
      description: Create a new golf
      parameters:
      - description: Golf
        in: body
        name: golf
        required: true
        schema:
          $ref: '#/definitions/Golf'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Golf'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Create a new golf
      tags:
      - Golfs
  /v1/golfs/{golf_id}/courses:
    get:
      parameters:
      - description: Golf ID
        in: path
        name: golf_id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Course'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Get all courses for a specific golf
      tags:
      - Courses
  /v1/golfs/{id}:
    delete:
      description: Delete an existing golf by ID
      parameters:
      - description: Golf ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Delete an existing golf
      tags:
      - Golfs
    get:
      description: Get a golf by ID
      parameters:
      - description: Golf ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Golf'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Get a golf by ID
      tags:
      - Golfs
    patch:
      consumes:
      - application/json
      description: Update an existing golf by ID
      parameters:
      - description: Golf ID
        in: path
        name: id
        required: true
        type: string
      - description: Golf
        in: body
        name: golf
        required: true
        schema:
          $ref: '#/definitions/Golf'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Golf'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Update an existing golf
      tags:
      - Golfs
  /v1/holes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Hole'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Get all holes
      tags:
      - Holes
    post:
      consumes:
      - application/json
      parameters:
      - description: Hole
        in: body
        name: hole
        required: true
        schema:
          $ref: '#/definitions/Hole'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Hole'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Create a new hole
      tags:
      - Holes
  /v1/holes/{id}:
    delete:
      parameters:
      - description: Hole ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Delete an existing hole
      tags:
      - Holes
    get:
      parameters:
      - description: Hole ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Hole'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Get a hole by ID
      tags:
      - Holes
    patch:
      consumes:
      - application/json
      parameters:
      - description: Hole ID
        in: path
        name: id
        required: true
        type: string
      - description: Hole
        in: body
        name: hole
        required: true
        schema:
          $ref: '#/definitions/Hole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Hole'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Update an existing hole
      tags:
      - Holes
  /v1/tees:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Tee'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Get all tees
      tags:
      - Tees
    post:
      consumes:
      - application/json
      parameters:
      - description: Tee
        in: body
        name: tee
        required: true
        schema:
          $ref: '#/definitions/Tee'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Tee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Create a new tee
      tags:
      - Tees
  /v1/tees/{id}:
    delete:
      parameters:
      - description: Tee ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Delete an existing tee
      tags:
      - Tees
    get:
      parameters:
      - description: Tee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Tee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Get a tee by ID
      tags:
      - Tees
    patch:
      consumes:
      - application/json
      parameters:
      - description: Tee ID
        in: path
        name: id
        required: true
        type: string
      - description: Tee
        in: body
        name: tee
        required: true
        schema:
          $ref: '#/definitions/Tee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Tee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AppError'
      summary: Update an existing tee
      tags:
      - Tees
swagger: "2.0"